requirements.txt
python-telegram-bot==13.15
requests==2.29.0
python-dotenv==1.0.0
urllib3==1.26.15